net: mqtt_sn: move library part to lib/mqtt_sn
- The library has no technical dependency on the network stack. The
  transport is fully abstracted via an API struct and only the UDP
  transport implementation needs a network stack.
- MQTT-SN has no specified dependency on the concept of a network. The
  specification doesn't describe any specific transport. The gateway
  discovery messages are the only ones which contain addresses, so anything
  besides that doesn't even need addresses or ports.
- This allows using the mqtt_sn library over custom transports (like UART)
  on devices with very little RAM and ROM. Of course, this needs a custom
  gateway on the other side to either translate to e.g. UDP or feed
  into an MQTT server directly.

The UDP transport is left within the network subsystem. I added a new
header for it and moved all of its definitions to it. Thus, the generic
mqtt_sn header doesn't contain anything related to the UDP transport
anymore.

With the netif dependency removed, the tests select SMP capable boards now.
Since the tests are written under the assumption, that threads don't run in
parallel, I added a filter to exclude these.
