+++ lib/CMakeLists.txt
@@ -9,6 +9,7 @@ add_subdirectory(hash)
 add_subdirectory(heap)
 add_subdirectory(mem_blocks)
 add_subdirectory(midi2)
+add_subdirectory_ifdef(CONFIG_MQTT_SN_LIB mqtt_sn)
 add_subdirectory_ifdef(CONFIG_NET_BUF net_buf)
 add_subdirectory(os)
 add_subdirectory(utils)
