+++ lib/mqtt_sn/CMakeLists.txt
@@ -3,5 +3,7 @@
 zephyr_library()
 
 zephyr_library_sources(
-  mqtt_sn_transport_udp.c
+  mqtt_sn.c
+  mqtt_sn_decoder.c
+  mqtt_sn_encoder.c
 )
