+++ samples/net/mqtt_sn_publisher/src/main.c
@@ -10,7 +10,7 @@
 #include <stdio.h>
 
 #include <zephyr/kernel.h>
-#include <zephyr/net/mqtt_sn.h>
+#include <zephyr/mqtt_sn.h>
 #include <zephyr/net/conn_mgr_monitor.h>
 #include <zephyr/net/net_mgmt.h>
 #include <zephyr/net/socket.h>
