+++ subsys/net/lib/CMakeLists.txt
@@ -8,7 +8,7 @@ add_subdirectory_ifdef(CONFIG_SOCKS socks)
 add_subdirectory_ifdef(CONFIG_SNTP sntp)
 add_subdirectory_ifdef(CONFIG_MCP_SERVER             mcp)
 add_subdirectory_ifdef(CONFIG_MQTT_LIB               mqtt)
-add_subdirectory_ifdef(CONFIG_MQTT_SN_LIB            mqtt_sn)
+add_subdirectory_ifdef(CONFIG_MQTT_SN_TRANSPORT_UDP  mqtt_sn)
 add_subdirectory_ifdef(CONFIG_PTP                    ptp)
 add_subdirectory_ifdef(CONFIG_RTP                    rtp)
 add_subdirectory_ifdef(CONFIG_NET_LATMON             latmon)
