+++ tests/lib/mqtt_sn_client/prj.conf
@@ -1,12 +1,7 @@
-# required for htons
-CONFIG_NETWORKING=y
-CONFIG_NET_TEST=y
-
 # native IP stack support
 CONFIG_ENTROPY_GENERATOR=y
 CONFIG_TEST_RANDOM_GENERATOR=y
 
-CONFIG_NET_LOG=y
 CONFIG_MQTT_SN_LOG_LEVEL_DBG=y
 
 # enable the MQTT lib
