+++ tests/lib/mqtt_sn_client/src/mqtt_sn_client.c
@@ -4,7 +4,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-#include <zephyr/net/mqtt_sn.h>
+#include <zephyr/mqtt_sn.h>
 #include <zephyr/sys/util.h> /* for ARRAY_SIZE */
 #include <zephyr/tc_util.h>
 #include <zephyr/ztest.h>
