+++ tests/lib/mqtt_sn_client/tests.yaml
@@ -1,0 +1,6 @@
+tests:
+  libraries.mqtt_sn.client:
+    min_ram: 16
+    tags:
+      - mqtt_sn
+    filter: not CONFIG_SMP
