+++ tests/lib/mqtt_sn_packet/CMakeLists.txt
@@ -6,8 +6,7 @@ find_package(Zephyr REQUIRED HINTS $ENV{ZEPHYR_BASE})
 project(mqtt_sn_packet)
 
 target_include_directories(app PRIVATE
-  ${ZEPHYR_BASE}/subsys/net/ip
-  ${ZEPHYR_BASE}/subsys/net/lib/mqtt_sn
+  ${ZEPHYR_BASE}/lib/mqtt_sn
 )
-FILE(GLOB app_sources src/*.c)
+file(GLOB app_sources src/*.c)
 target_sources(app PRIVATE ${app_sources})
