+++ tests/lib/mqtt_sn_packet/prj.conf
