+++ tests/lib/mqtt_sn_packet/tests.yaml
@@ -1,6 +1,5 @@
 tests:
-  net.mqtt_sn.packet:
+  libraries.mqtt_sn.packet:
     min_ram: 16
     tags:
       - mqtt_sn
-      - net
