--- tests/net/lib/mqtt_sn_client/tests.yaml
@@ -1,8 +1,0 @@
-common:
-  depends_on: netif
-tests:
-  net.mqtt_sn.client:
-    min_ram: 16
-    tags:
-      - mqtt_sn
-      - net
